+++ src/Facades/InvestorsExchangeData/Stats/Recent.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Digitonic\IexCloudSdk\Facades\InvestorsExchangeData\Stats;
+
+use Illuminate\Support\Facades\Facade;
+
+class Recent extends Facade
+{
+    /**
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return \Digitonic\IexCloudSdk\InvestorsExchangeData\Stats\Recent::class;
+    }
+}